requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.3.11=hca03da5_0
certifi=2024.6.2=pypi_0
distro=1.9.0=pypi_0
func-timeout=4.3.5=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=pypi_0
libffi=3.4.4=hca03da5_1
ncurses=6.4=h313beb8_0
numpy=2.0.0=pypi_0
openai=1.34.0=pypi_0
openssl=3.0.14=h80987f9_0
pip=24.0=py311hca03da5_0
psycopg2-binary=2.9.9=pypi_0
pydantic=2.7.4=pypi_0
pydantic-core=2.18.4=pypi_0
pymysql=1.1.1=pypi_0
python=3.11.5=hb885b13_0
readline=8.2=h1a28f6b_0
setuptools=69.5.1=py311hca03da5_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h80987f9_0
tk=8.6.14=h6ba3021_0
tqdm=4.66.4=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
wheel=0.43.0=py311hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1

boto3~=1.37.22
psycopg2-binary~=2.9.10
sql_metadata~=2.15.0
nltk~=3.9.1